donation all

+++ Models/donations.cs
@@ -4,7 +4,7 @@ namespace API_GroupProject.Models
     {
         public string Email {get; set;}
         public string DonorName {get; set;}
-        public double MoneyDonated {get; set;}
+        public int MoneyDonated {get; set;}
         public string Date {get; set;}
     }
 }
